backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
duckdb==1.3.2
pandas==2.3.2
pyarrow==21.0.0
python-dotenv==1.0.1
